moving ErrorBoundary to ui/components
